+++ implementation/cdi/src/main/java/io/smallrye/opentelemetry/implementation/config/traces/jaeger/JaegerExporterConfig.java
@@ -1,11 +1,14 @@
 package io.smallrye.opentelemetry.implementation.config.traces.jaeger;
 
 import io.smallrye.config.ConfigMapping;
+import io.smallrye.config.WithDefault;
 import io.smallrye.opentelemetry.implementation.config.ConnectionConfig;
+import io.smallrye.opentelemetry.implementation.config.traces.otlp.ExporterOtlpTraceConfig;
+
+import java.util.Optional;
 
 @ConfigMapping(prefix = "otel.exporter.jaeger")
-public interface ExporterJaegerTraceConfig extends ConnectionConfig {
-    // TODO default Jaeger endpoint
+public interface JaegerExporterConfig extends ConnectionConfig {
     @Override
     String endpoint();
 
@@ -18,4 +21,7 @@ public interface ExporterJaegerTraceConfig extends ConnectionConfig {
     //    String timeout(); from parent
 
     //    ExporterType protocol(); from parent
+    ExporterJaegerTraceConfig traces();
+    // TODO metrics();
+    // TODO logs();
 }
