+++ implementation/cdi/src/main/java/io/smallrye/opentelemetry/implementation/config/traces/otlp/ExporterOtlpTraceConfig.java
@@ -1,10 +1,10 @@
 package io.smallrye.opentelemetry.implementation.config.traces.otlp;
 
 import io.smallrye.config.ConfigMapping;
-import io.smallrye.opentelemetry.implementation.config.OtelConnectionConfig;
+import io.smallrye.opentelemetry.implementation.config.ConnectionConfig;
 
 @ConfigMapping(prefix = "otel.exporter.otlp")
-public interface ExporterOtlpTraceConfig extends OtelConnectionConfig {
+public interface ExporterOtlpTraceConfig extends ConnectionConfig {
 
     /**
      * Sets the OTLP endpoint to connect to. If unset, defaults to {@value Constants#DEFAULT_ENDPOINT_URL}.
